version: Move semver to internal package
This moves the semver types to an internal package so this can be re-used
elsewhere.
